+++ lib/oauth/tty.rb
@@ -1,9 +1,11 @@
 # frozen_string_literal: true
 
 # stdlib
+require "cgi"
 require "optparse"
 
 # external gems
+require "auth/sanitizer"
 require "version_gem"
 
 # For initial release as a standalone gem, this gem must not declare oauth as a dependency,